Enhance: use NamingEvent instead of ServiceInfo
- remove unnecessary error enum
