+++ src/naming/dto/mod.rs
@@ -1,0 +1,2 @@
+mod service_info;
+pub(crate) use service_info::*;
