+++ src/naming/message/response/subscribe_service_response.rs
@@ -1,6 +1,6 @@
 use nacos_macro::response;
 
-use crate::api::naming::ServiceInfo;
+use crate::naming::dto::ServiceInfo;
 
 #[response(identity = "SubscribeServiceResponse", module = "naming")]
 pub(crate) struct SubscribeServiceResponse {
